update for use swagger

+++ RestAPIWithASP-Net5/RestAPIWithASP-Net5/Business/IPersonBusiness.cs
@@ -6,9 +6,9 @@ namespace RestAPIWithASP_Net5.person.Business
     public interface IPersonBusiness
     {
         Person Create(Person person);
-        Person FindById(long id);
+        Person FindById(int id);
         List<Person> FindAll();
         Person Update(Person person);
-        void Delete(long id);
+        void Delete(int id);
     }
 }
